Refactor Login.js, split out Styles into own file, use es7 imports, update npm dependencies and tweak webpack-config
